Guard temp file removal in finally block with existence check
Agent-Logs-Url: https://github.com/sensepost/objection/sessions/38e164a4-354a-4ca7-9a88-73084ba7fa00

+++ objection/utils/patchers/ios.py
@@ -481,7 +481,8 @@ class IosPatcher(BasePlatformPatcher):
 
         finally:
             # cleanup the temp path
-            os.remove(decoded_location)
+            if os.path.exists(decoded_location):
+                os.remove(decoded_location)
 
         click.secho('Mobile provision bundle identifier is: {}'.format(self.bundle_id), dim=True)
